requirements.txt
pystray
bleak
Pillow
win10toast